+++ client/pages/index.tsx
@@ -1,5 +1,9 @@
-const Home: React.FC = () => {
-  return <div>Hello</div>;
+import type { NextPage } from "next";
+import Head from "next/head";
+import Image from "next/image";
+
+const Home: NextPage = () => {
+  return <div className="font-bold text-3xl">Hello next</div>;
 };
 
 export default Home;
